requirements.txt
-c constraints.txt
bumpversion
flake8
ipython
jsonschema
lxml
python-dateutil
pytest
pytest-cov
pytest-mock
rpc-zigzag
setuptools
sh
testinfra
tox
twine
